// backend-cm/prisma/investigation/generated/ticket_type_code.ts
import { ticket } from "./ticket";
import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";

export class ticket_type_code {
  @ApiProperty({ type: String })
  ticket_type_code: string;

  @ApiProperty({ type: String })
  short_description: string;

  @ApiPropertyOptional({ type: String })
  long_description?: string;

  @ApiPropertyOptional({ type: Number })
  display_order?: number;

  @ApiProperty({ type: Boolean })
  active_ind: boolean = true;

  @ApiProperty({ type: String })
  create_user_id: string;

  @ApiProperty({ type: Date })
  create_utc_timestamp: Date;

  @ApiPropertyOptional({ type: String })
  update_user_id?: string;

  @ApiPropertyOptional({ type: Date })
  update_utc_timestamp?: Date;

  @ApiProperty({ isArray: true, type: () => ticket })
  ticket_ticket_ticket_type_codeToticket_type_code: ticket[];
}